generator/requirements.txt
google-cloud-aiplatform
google-generativeai
supabase
GitPython
tenacity
asyncio
python-dotenv
tqdm
pydantic
